cleaned up the files and started the event finder viewbar
