+++ public/index.html
@@ -57,5 +57,6 @@
     -->
  </div>
     <script>var mountNode = document.getElementById('container');</script>
+  
   </body>
 </html>
