+++ src/App.js
@@ -1,27 +1,31 @@
 import React, { Component } from 'react';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import './App.css';
-import 'antd-mobile/dist/antd-mobile.css';
-
-import 'antd-mobile/lib/date-picker/style/css';
+import { BrowserRouter, Route, Switch} from 'react-router-dom';
 import Home from './pages/Home';
-import logIn from './pages/LogIn';
-import Navigation from './components/Navigation';
+import LogIn from './pages/LogIn';
 import CreateAccount from './pages/CreateAccount';
+import Navigation from './components/NavBar';
 import Footer from './components/Footer';
-// import Menu from './components/NavBar';
 
-// import NavBar from './components/NavBar';
+import 'antd-mobile/dist/antd-mobile.css';
+import 'antd-mobile/lib/date-picker/style/css';
+import './App.css';
+// import './index';
+
+
+
+
+
+
 class App extends Component {
   render() {
     return (
       <BrowserRouter>
         <div>
-          <Navigation />
+          <Navigation/>
           <Switch>
-            <Route exact path="/" component={Home}/>
-            <Route exact path="/logIn" component={logIn} />
-            <Route exact path = "/CreateAccount" component={CreateAccount}/>
+            <Route exact path ="/CreateAccount" component={CreateAccount}/>
+            <Route exact path="/Home" component={Home}/>
+            <Route exact path="LogIn" component={LogIn} />
           </Switch>
           <Footer/>
         </div>
