+++ src/components/LogInForm.js
@@ -1,0 +1,80 @@
+import React from 'react';
+import {LINK} from 'react';
+import { Form, Input, Button, Checkbox} from 'antd';
+import { UserOutlined, LockOutlined } from '@ant-design/icons';
+
+import "../index.css";
+
+
+const LogInForm = () => {
+
+  const styles = {
+    formBorder: {
+    margin: `${200}px`,
+    padding: `${100}px`,
+    border: `${2}px`,
+    borderColor: `black`,
+    borderStyle: `solid`,
+  }
+  };
+
+    return (
+      <div className="container">
+        <div className="align">
+        <Form style={styles.formBorder}
+      name="normal_login"
+      className="login-form"
+      
+      initialValues={{
+        remember: true,
+      }}
+    >
+      <Form.Item
+        name="username"
+        rules={[
+          {
+            required: true,
+            message: 'Please input your Username!',
+          },
+        ]}
+      >
+        <Input prefix={<UserOutlined className="site-form-item-icon" />} placeholder="Username" />
+      </Form.Item>
+      <Form.Item
+        name="password"
+        rules={[
+          {
+            required: true,
+            message: 'Please input your Password!',
+          },
+        ]}
+      >
+        <Input
+          prefix={<LockOutlined className="site-form-item-icon" />}
+          type="password"
+          placeholder="Password"
+        />
+      </Form.Item>
+      <Form.Item>
+        <Form.Item name="remember" valuePropName="checked" noStyle>
+          <Checkbox>Remember me</Checkbox>
+        </Form.Item>
+
+        <LINK className="login-form-forgot" href="">
+          Forgot password
+        </LINK>
+      </Form.Item>
+
+      <Form.Item>
+        <Button type="primary" htmlType="submit" className="login-form-button">
+          Log in
+        </Button>
+        Or <LINK href="">register now!</LINK>
+      </Form.Item>
+    </Form>
+  
+    </div>
+        </div>
+    );
+}
+export default LogInForm;
