+++ src/components/NavBar.js
@@ -1,34 +1,39 @@
-// import React from 'react';
-// import { Tabs, TabBar  } from 'antd-mobile';
-// import '../index';
-// import Navigation from '../components/Navigation';
+import React from "react";
+import {NavLink} from "react-router-dom";
+import { NavBar, Icon} from "antd-mobile";
 
-// export default function () {
-//     return (
-//         <div>
-// <div className="Container">
-//             <div className="align">
-//                 <NavBar>
-//                     <Navigation/>
-//                     <TabBar>
-//                         <Tabs.Item>
-//                             Home
-//                         </Tabs.Item>
-//                         <Tabs.Item>
-//                             Log In
-//                         </Tabs.Item>
-//                         <Tabs.Item>
-//                             My Account
-//                         </Tabs.Item>
-//                     </TabBar>
-//                 </NavBar>
-//             </div>
-//         </div>
-//         </div>
-//     )
-// }
+import "../index";
 
+const Navigation = () => {
+  const styles = {
+    navStyle: {
+      display: "flex",
+      justifyContent: "space-around",
+      color: "White",
+      backgroundColor: "blueviolet",
+      padding: `${50}px`,
+    },
+  };
 
+  return (
+    <>
+      <NavBar
+        className="container"
+        style={styles.navStyle}
+        mode="light"
+        leftContent={[<Icon key="1" type="ellipsis" />]}
+        rightContent={"home"}
+      >
+        <h1>GatherNow</h1>
+      </NavBar>
+      <div>
+      <NavLink to="/CreateAccount">CreateAccount</NavLink>,
+        <NavLink to="/Home">Home</NavLink>,
+        <NavLink to="/LogIn">LogIn</NavLink>,
+      </div>
+      <div></div>
+    </>
+  );
+};
 
-
-
+export default Navigation;
