+++ src/pages/LogIn.js
@@ -1,79 +1,13 @@
-import React from 'react';
-import { Form, Input, Button, Checkbox } from 'antd';
-import { UserOutlined, LockOutlined } from '@ant-design/icons';
-import "../index.css"
+import React from "react";
+import LogInForm from "../components/LogInForm";
 
+import '../index';
 
-const LogIn = () => {
-
-  const styles = {
-    formBorder: {
-    margin: `${200}px`,
-    padding: `${100}px`,
-    border: `${2}px`,
-    borderColor: `black`,
-    borderStyle: `solid`,
-  }
-  }
-
-    return (
-      <div className="container">
-        <div className="align">
-        <Form style={styles.formBorder}
-      name="normal_login"
-      className="login-form"
-      
-      initialValues={{
-        remember: true,
-      }}
-    >
-      <Form.Item
-        name="username"
-        rules={[
-          {
-            required: true,
-            message: 'Please input your Username!',
-          },
-        ]}
-      >
-        <Input prefix={<UserOutlined className="site-form-item-icon" />} placeholder="Username" />
-      </Form.Item>
-      <Form.Item
-        name="password"
-        rules={[
-          {
-            required: true,
-            message: 'Please input your Password!',
-          },
-        ]}
-      >
-        <Input
-          prefix={<LockOutlined className="site-form-item-icon" />}
-          type="password"
-          placeholder="Password"
-        />
-      </Form.Item>
-      <Form.Item>
-        <Form.Item name="remember" valuePropName="checked" noStyle>
-          <Checkbox>Remember me</Checkbox>
-        </Form.Item>
-
-        <a className="login-form-forgot" href="">
-          Forgot password
-        </a>
-      </Form.Item>
-
-      <Form.Item>
-        <Button type="primary" htmlType="submit" className="login-form-button">
-          Log in
-        </Button>
-        Or <a href="">register now!</a>
-      </Form.Item>
-    </Form>
+export default function LogIn() {
+  return(
+    <>
+    <LogInForm/>
   
-    </div>
-        </div>
-    );
+</>
+  )
 }
-
-export default LogIn;
